feat: update to typescript
