+++ SectionListBasics.tsx
@@ -1,0 +1,153 @@
+import * as React from "react";
+import { SectionList, StyleSheet, Text, TextInput, View } from "react-native";
+
+import { pluralize } from "./helpers/stringHelper";
+import { SelectedLettersText } from "./SelectedLettersText";
+
+/*
+  React-native does not support dynamic imports apparently
+  See https://github.com/facebook/react-native/issues/2481#issuecomment-299074154
+*/
+const DATA: any = {
+  letterCountIs2: require("./data/words/2-letters.json"),
+  letterCountIs3: require("./data/words/3-letters.json")
+};
+
+export interface SectionListBasicsProps {
+  letterCount?: number;
+}
+
+const SectionListBasics: React.FC<SectionListBasicsProps> = props => {
+  // Hooks
+  const [searchText, setSearchText] = React.useState("");
+  const [isSearch, setIsSearch] = React.useState(false);
+
+  const { letterCount } = props;
+
+  const getXLetterWordsSections = (jsonObject: any) => {
+    const allMixedUpWords = isSearch
+      ? jsonObject.filter((wordObject: any) =>
+          wordObject.englishWord.toLowerCase().includes(searchText)
+        )
+      : jsonObject;
+    const allSortedWords: any[] = [];
+    let currentGroupOfWords: any = [];
+    let currentLetter = 97;
+
+    allMixedUpWords.forEach((wordObject: any) => {
+      const firstEnglishLetter = wordObject.englishWord[0].toLowerCase();
+      if (firstEnglishLetter === String.fromCharCode(currentLetter)) {
+        currentGroupOfWords.push(wordObject.englishWord);
+      } else {
+        if (currentGroupOfWords.length > 0) {
+          allSortedWords.push(currentGroupOfWords);
+        }
+        currentGroupOfWords = [wordObject.englishWord];
+        currentLetter = firstEnglishLetter.charCodeAt(0);
+      }
+    });
+    if (currentGroupOfWords.length > 0) {
+      allSortedWords.push(currentGroupOfWords);
+    }
+
+    return allSortedWords.map(groupOfWords => {
+      const letterSection = groupOfWords[0][0].toUpperCase();
+      return {
+        title: `${letterSection} (${pluralize("word", groupOfWords.length)})`,
+        data: groupOfWords
+      };
+    });
+  };
+
+  const key = `letterCountIs${letterCount}`;
+  const jsonObject = DATA[key];
+  const xLetterWordsSections = getXLetterWordsSections(jsonObject);
+  const occurrencesLetters = isSearch ? searchText : undefined;
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>{`${letterCount}-letter words`}</Text>
+      <TextInput
+        placeholder="Search for a word..."
+        onChangeText={text => {
+          const trimmedText = text.trim();
+          if (trimmedText.length > 0) {
+            setSearchText(trimmedText.toLowerCase());
+            setIsSearch(true);
+          } else {
+            setIsSearch(false);
+          }
+        }}
+        style={styles.searchBar}
+      />
+      {xLetterWordsSections.length ? (
+        <SectionList
+          sections={xLetterWordsSections}
+          renderItem={({ item }) => (
+            <SelectedLettersText
+              letters={occurrencesLetters}
+              style={styles.item}
+              text={item}
+            />
+          )}
+          renderSectionHeader={({ section }) => (
+            <Text style={styles.sectionHeader}>{section.title}</Text>
+          )}
+          keyExtractor={(_item, index) => index.toString()}
+        />
+      ) : (
+        <Text style={styles.noResult}>
+          {`No word was found for your search.${
+            isSearch && letterCount && searchText.length > letterCount
+              ? ` This view is for ${letterCount}-letter words only.`
+              : ""
+          }`}
+        </Text>
+      )}
+    </View>
+  );
+};
+
+SectionListBasics.defaultProps = {
+  letterCount: 2
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    paddingTop: 22
+  },
+  searchBar: {
+    fontSize: 16,
+    borderColor: "black",
+    textAlign: "left",
+    paddingLeft: 10,
+    paddingBottom: 10
+  },
+  noResult: {
+    paddingLeft: 10
+  },
+  title: {
+    fontSize: 20,
+    fontWeight: "bold",
+    paddingBottom: 10,
+    textAlign: "center"
+  },
+  sectionHeader: {
+    color: "white",
+    paddingTop: 2,
+    paddingLeft: 10,
+    paddingRight: 10,
+    paddingBottom: 2,
+    fontSize: 14,
+    fontWeight: "bold",
+    backgroundColor: "black"
+  },
+  item: {
+    padding: 10,
+    fontSize: 18,
+    height: 44
+  }
+});
+
+export { SectionListBasics };
