--- helpers/stringHelper.js
@@ -1,6 +1,0 @@
-const pluralize = (word, count) =>
-	count === 1
-		? `${count} ${word}`
-		: `${count} ${word}s`;
-
-export { pluralize };
