+++ helpers/stringHelper.ts
@@ -1,0 +1,4 @@
+const pluralize = (word: string, count: number) =>
+  count === 1 ? `${count} ${word}` : `${count} ${word}s`;
+
+export { pluralize };
