+++ jest.config.js
@@ -1,5 +1,6 @@
 module.exports = {
-  preset: "jest-expo",
+  moduleFileExtensions: ["js", "json", "jsx", "node", "ts", "tsx"],
+  preset: "react-native",
   transform: {
     "\\.js$": "<rootDir>/node_modules/react-native/jest/preprocessor.js"
   }
