+++ package.json
@@ -3,6 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
+    "@types/jest": "^24.0.21",
+    "@types/react": "^16.9.11",
+    "@types/react-native": "^0.60.22",
+    "@types/react-test-renderer": "^16.9.1",
     "babel-preset-expo": "^7.1.0",
     "csv-parse": "^4.6.5",
     "expo-cli": "^3.4.1",
